+++ test/augmenters/test_blend.py
@@ -752,21 +752,28 @@ class TestAlpha(unittest.TestCase):
             iaa.Noop(),
             iaa.Affine(translate_px={"x": 1}),
             per_channel=True)
-        expected_same = self.psoi.deepcopy()
-        expected_shifted = self.psoi.shift(left=1)
-        seen = [0, 0]
+        psoi = self.psoi
+        expected_same = psoi.deepcopy()
+        expected_shifted = psoi.shift(left=1)
+        seen = [0, 0, 0]
         for _ in sm.xrange(200):
-            observed = aug.augment_polygons([self.psoi])[0]
-            if observed.polygons[0].exterior_almost_equals(
-                    expected_same.polygons[0]):
+            observed = aug.augment_polygons([psoi])[0]
+            # We use here allclose() instead of exterior_almost_equals()
+            # as the latter one is much slower and we don't have to deal
+            # with tricky geometry changes here, just naive shifting.
+            if np.allclose(observed.polygons[0].exterior,
+                           expected_same.polygons[0].exterior,
+                           rtol=0, atol=0.1):
                 seen[0] += 1
-            elif observed.polygons[0].exterior_almost_equals(
-                    expected_shifted.polygons[0]):
+            elif np.allclose(observed.polygons[0].exterior,
+                             expected_shifted.polygons[0].exterior,
+                             rtol=0, atol=0.1):
                 seen[1] += 1
             else:
-                assert False
+                seen[2] += 1
         assert 100 - 50 < seen[0] < 100 + 50
         assert 100 - 50 < seen[1] < 100 + 50
+        assert seen[2] == 0
 
     def test_empty_polygons(self):
         # empty polygons
@@ -1311,15 +1318,24 @@ class TestAlphaElementwise(unittest.TestCase):
             else:
                 seen[2] += 1
 
-        nb_points = len(ps[0].exterior)
-        p_all_same = 2 * ((1/2)**nb_points)  # all points moved in the same way
-        expected_iter = nb_iterations*p_all_same
-        expected_iter_notsame = nb_iterations*(1-p_all_same)
-        atol = nb_iterations * (5*p_all_same)
+        # This code can be used if the polygon augmentation mode is
+        # AlphaElementwise._MODE_POINTWISE. Currently it is _MODE_EITHER_OR.
+        #
+        # nb_points = len(ps[0].exterior)
+        # p_all_same = 2 * ((1/2)**nb_points)  # all points moved in same way
+        # expected_iter = nb_iterations*p_all_same
+        # expected_iter_notsame = nb_iterations*(1-p_all_same)
+        # atol = nb_iterations * (5*p_all_same)
+        #
+        # assert np.isclose(seen[0], expected_iter, rtol=0, atol=atol)
+        # assert np.isclose(seen[1], expected_iter, rtol=0, atol=atol)
+        # assert np.isclose(seen[2], expected_iter_notsame, rtol=0, atol=atol)
 
+        expected_iter = nb_iterations*0.5
+        atol = nb_iterations*0.15
         assert np.isclose(seen[0], expected_iter, rtol=0, atol=atol)
         assert np.isclose(seen[1], expected_iter, rtol=0, atol=atol)
-        assert np.isclose(seen[2], expected_iter_notsame, rtol=0, atol=atol)
+        assert seen[2] == 0
 
     def test_empty_polygons(self):
         psoi = ia.PolygonsOnImage([], shape=(1, 2, 3))